Create README.md
"Hello! I'm Sunayana Chowdary Vunnam, a soon-to-be Master's graduate in Biotechnology with a passion for harnessing the power of Artificial Intelligence and Machine Learning (AIML) in biological research. With a strong foundation in biotechnology and a growing expertise in AIML, I'm excited to explore the intersection of technology and biology. I'm currently learning and applying AIML concepts to drive innovation in the field of biotechnology. Welcome to my GitHub profile, where I'll be sharing my projects, experiments, and learnings as I navigate the fascinating world of bioinformatics and computational biology!"